pyproject.toml
[tool.poetry]
name = "online-payments"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
django-cors-headers = "^4.7.0"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.4.0"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^8.0.0"}
django-filter = "^25.1"
redis = "^5.2.1"
celery = "^5.4.0"
gevent = "^24.11.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-django = "^4.10.0"
pytest-cov = "^6.0.0"
ipython = "^8.32.0"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.2"
black = "^25.1.0"
isort = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
